#orderbook change type quantity to decimal

+++ src/BitstampOrderBookService/Presentation/Controllers/PricingController.cs
@@ -15,7 +15,7 @@ namespace BitstampOrderBookService.Presentation.Controllers
         }
 
         [HttpGet("simulate")]
-        public async Task<IActionResult> SimulateBestPrice(string operation, string instrument, int quantity)
+        public async Task<IActionResult> SimulateBestPrice(string operation, string instrument, decimal quantity)
         {
             try
             {
